fix(editor): stop pastes throwing on non-global iframe paste rules
The iframe embed paste rules used non-global regexes. TipTap's paste-rule
engine runs each `find` through String.prototype.matchAll(), which throws
"called with a non-global RegExp argument" — so every paste into the editor
blew up in appendTransaction and the iframe paste rules never worked.

Add the `g` flag to both patterns. The HTML pattern keeps a non-global form
for its .exec() parsing path (a global regex would carry lastIndex across
calls) and exposes a global twin for the paste rule.

Extract the pure embed-parsing helpers into iframe-embed.js so they can be
unit-tested without the Vue NodeView, which also breaks the
iframe-block.js <-> IframeBlockView.vue import cycle.
